+++ Coditech.Project/Admin/Coditech.Admin/ViewModel/DataTableViewModel.cs
@@ -10,7 +10,7 @@ namespace Coditech.Admin.ViewModel
         public int PageIndex { get; set; } = 1;
         public int PageSize { get; set; } = 10;
         public string SelectedCentreCode { get; set; } = string.Empty;
-        public int SelectedDepartmentId { get; set; }
+        public short SelectedDepartmentId { get; set; }
         public string SelectedParameter1 { get; set; } = string.Empty;
         public string SelectedParameter2 { get; set; } = string.Empty;
         public string SelectedParameter3 { get; set; } = string.Empty;
